requirements.txt
tensorflow>=2.10.0
streamlit>=1.15.0
requests>=2.28.0
pillow>=9.0.0
numpy>=1.23.0
opencv-python-headless

# dataset
# https://www.kaggle.com/datasets/techsash/waste-classification-data